Lots of Update (Read Description)
Users can now input Pictures in Gallery, the images will be saved in the gallery folder, and the path will be saved in the new article_gallery table.

The tags are also now functional too, the writers can now assign different tags to their own articles.

Both tags and gallery can now be seen in the reviewer side and the lundayan site page dynamically.