Lots of Updates Again (Read Description)
Lundayan Site Home is now only showing approved and posted articles within scheduled post date and expiration date.

Articles won't show up on home lundayan page if they dont have post date and expiration date, the same goes if they are not approved and is only a draft or not published.

Revamped the edit article in the writer side to display 2 input date type boxes what will serve as the post date and expiration date of the article.

Both will have a red border if not set properly, if they are both valid and are both not null, they will be set as the post date and expiration date of the article.

The post button is updated to be 'Publish Article' and 'Unpublish Article' giving the writer the power to retract his posted articles if he wants to hide it from the public or something.