Article, Calendar, Announcements
Lundayan Article is revamped a little bit, I am planning to add the author's credential on the side of the article information.

The Calendar now works, it highlights every day that has an article posted. Click the cell date would open a dialog, that contains the articles, they are clickable and can navigate users to the actual article site.

Announcements are now dynamic! I added article_type attribute in the articles table to determine if an article is regular or an announcement. The writer could also specify if he/she is writing an announcement or just a plain old regular article.

Been experimenting on using .htaccess for share-able link. I am planning to move everything to htdocs to to this feature... Please don't break :(