About, Lundayan Article, Lundayan Footer, Lundayan Upper Nav and Nav bar or Editor Side
About Page is now available! Users can now view what's all about Lundayan and it's publishers.

Users can now access the editor page through the about section, there will be an anchor link that asks if they are a member. This will take them to the sign in page, in which they could sign in to get to editor page.

The Lundayan Article got a revamp in UI and UX.

The Next and Previous Article buttons are now working!

Added extra detail about the article's writer and reviewers on the side of article content. The author and reviewer details will stick to the side of the page until it reaches the gallery section.

The footer and upper header part of the lundayan site is updated.

The upper part just gain 2 logos lundayan logo on the left and plp logo on the right.

The footer is also the same, it gained two logo images and at the same time, a brief message about who developed the site.

The Navigation Bar on the Editor Side gained a new item, the Lundayan Home. This takes the user back to the Lundayan Home Page and takes along the user_id from the session along so that he does not need to sign in again unless he signed out from the editor side.